requirements.txt
kagglehub
numpy
pandas
ipykernel
scikit-learn
langchain[openai, ollama]
pywaves
python-dotenv